Cap::ProfileIdRewriter - moved from lib/cap_profile_id_rewriter.rb
